// frontend/src/Utils/types/api.schemas.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * API
 * API for an Obsidian integrated TODO app
 * OpenAPI spec version: 1.0.0
 */
export interface Grade {
  readonly id: number;
  /** @maxLength 100 */
  name: string;
  percentage: number;
  scored: number;
  module_info_id: number;
}

export interface ModuleInfo {
  readonly primary_tag: PrimaryTag;
  description?: string;
  readonly grades: readonly Grade[];
  readonly sections: readonly Section[];
}

export interface Note {
  readonly id: number;
  /** @maxLength 255 */
  name: string;
  description?: string;
  date: string;
  completed?: boolean;
  readonly primary_tag: PrimaryTag;
  primary_tag_id: number;
  readonly subtags: readonly SubTag[];
  subtags_ids: number[];
  readonly urls: readonly NoteURL[];
  urls_ids: number[];
}

export interface NoteURL {
  readonly id: number;
  /** @maxLength 255 */
  alias: string;
  /** @maxLength 200 */
  url: string;
}

export interface PatchedGrade {
  readonly id?: number;
  /** @maxLength 100 */
  name?: string;
  percentage?: number;
  scored?: number;
  module_info_id?: number;
}

export interface PatchedModuleInfo {
  readonly primary_tag?: PrimaryTag;
  description?: string;
  readonly grades?: readonly Grade[];
  readonly sections?: readonly Section[];
}

export interface PatchedNote {
  readonly id?: number;
  /** @maxLength 255 */
  name?: string;
  description?: string;
  date?: string;
  completed?: boolean;
  readonly primary_tag?: PrimaryTag;
  primary_tag_id?: number;
  readonly subtags?: readonly SubTag[];
  subtags_ids?: number[];
  readonly urls?: readonly NoteURL[];
  urls_ids?: number[];
}

export interface PatchedNoteURL {
  readonly id?: number;
  /** @maxLength 255 */
  alias?: string;
  /** @maxLength 200 */
  url?: string;
}

export interface PatchedPrimaryTag {
  readonly id?: number;
  /** @maxLength 100 */
  name?: string;
  /** @maxLength 7 */
  color?: string;
  readonly subtags?: readonly SubTag[];
}

export interface PatchedSection {
  readonly id?: number;
  readonly subtag?: SubTag;
  subtag_id?: number;
  module_info_id?: number;
  readonly notes?: string;
  readonly median_date?: string;
}

export interface PatchedSubTag {
  readonly id?: number;
  /** @maxLength 100 */
  name?: string;
  parent?: number;
}

export interface PrimaryTag {
  readonly id: number;
  /** @maxLength 100 */
  name: string;
  /** @maxLength 7 */
  color: string;
  readonly subtags: readonly SubTag[];
}

export interface Section {
  readonly id: number;
  readonly subtag: SubTag;
  subtag_id: number;
  module_info_id: number;
  readonly notes: string;
  readonly median_date: string;
}

export interface SubTag {
  readonly id: number;
  /** @maxLength 100 */
  name: string;
  parent: number;
}